Changed default file to decode, added notes on images and results, added some try catches on common exception points and attempted to "workaround" them

+++ src/JpegHelpers/DCT3.java
@@ -52,7 +52,7 @@ public class DCT3 {
                         );
                     }
                 }
-                matrix[y][x] = (int) Math.floorDiv(s, 4);
+                matrix[y][x] = Math.floorDiv(s, 4);
             }
         }
         return matrix;
